requirements.txt
flask==2.3.2
flasgger==0.9.5
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
joblib==1.3.2
pyyaml==6.0